+++ src/mailman/rest/tests/test_moderation.py
@@ -280,13 +280,13 @@ Hello
         self.assertEqual(len(items), 2)
         message = items[0].msg
         # First message is reject message to the sender.
-        self.assertEqual(message['From'], 'ant-bounces@example.com')
+        self.assertEqual(message['From'], 'ant-owner@example.com')
         self.assertEqual(message['To'], 'aperson@example.com')
         self.assertEqual(str(message['Subject']),
                          'Request to mailing list "Ant" rejected')
         fwd_msg = items[1].msg
         # Second message is the forwarded message to multiple recipients.
-        self.assertEqual(fwd_msg['From'], 'ant-bounces@example.com')
+        self.assertEqual(fwd_msg['From'], 'ant-owner@example.com')
         self.assertEqual(fwd_msg['To'], 'bee@example.com, me@example.com')
         self.assertEqual(str(fwd_msg['Subject']),
                          'Forward of moderated message')
@@ -313,7 +313,7 @@ Hello
         self.assertEqual(len(items), 1)
         fwd_msg = items[0].msg
         # Second message is the forwarded message.
-        self.assertEqual(fwd_msg['From'], 'ant-bounces@example.com')
+        self.assertEqual(fwd_msg['From'], 'ant-owner@example.com')
         self.assertEqual(fwd_msg['To'], 'bee@example.com')
         self.assertEqual(str(fwd_msg['Subject']),
                          'Forward of moderated message')
@@ -655,7 +655,7 @@ class TestSubscriptionModeration(unittest.TestCase):
         items = get_queue_messages('virgin')
         self.assertEqual(len(items), 1)
         message = items[0].msg
-        self.assertEqual(message['From'], 'ant-bounces@example.com')
+        self.assertEqual(message['From'], 'ant-owner@example.com')
         self.assertEqual(message['To'], 'anne@example.com')
         self.assertEqual(message['Subject'],
                          'Request to mailing list "Ant" rejected')
@@ -689,7 +689,7 @@ class TestSubscriptionModeration(unittest.TestCase):
         items = get_queue_messages('virgin')
         self.assertEqual(len(items), 1)
         message = items[0].msg
-        self.assertEqual(message['From'], 'ant-bounces@example.com')
+        self.assertEqual(message['From'], 'ant-owner@example.com')
         self.assertEqual(message['To'], 'anne@example.com')
         self.assertEqual(message['Subject'],
                          'Request to mailing list "Ant" rejected')